IncusCLI.is_instance_booted: accept "degraded" state
If any service fails to start, systemd will report the system in
"degraded" status. In that case, the system *is* booted, so we take it.
Also, if we one can't get in, then it's difficult to debug the actual
problem.

When reporting "degraded", systemctl will return non-zero, so we need to
allow failures when calling it.